@@ -3,7 +3,7 @@ package com.lacolinares.countdownnewyear;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.yasic.library.particletextview.MovingStrategy.RandomMovingStrategy;
 import com.yasic.library.particletextview.Object.ParticleTextViewConfig;
